pyproject.toml
[project]
name = "rlhf-poisoning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9,<3.13"
dependencies = [
    "datasets==2.20.0",
    "gurobipy>=12.0.2",
    "hydra-core==1.3.2",
    "transformers==4.45.2",
    "torch==2.5.1",
    "trl==0.11.4",
    "scipy==1.13.1",
    "peft==0.13.0",
]
